Move sharedpreferences operations to TasksData object.
